* Fix PHP 8 no longer interpreting empty string as 0

+++ public_html/includes/library/lib_length.inc.php
@@ -61,7 +61,7 @@
 
     public static function convert($value, $from, $to) {
 
-      if ($value == 0) return 0;
+      if ((float)$value == 0) return 0;
 
       if ($from == $to) return (float)$value;
 
@@ -75,7 +75,7 @@
         return;
       }
 
-      if (self::$classes[$from]['value'] == 0 || self::$classes[$to]['value'] == 0) return;
+      if ((float)self::$classes[$from]['value'] == 0 || (float)self::$classes[$to]['value'] == 0) return;
 
       return $value * (self::$classes[$to]['value'] / self::$classes[$from]['value']);
     }
